requirements.txt
# Core web framework and server
fastapi>=0.104.1
uvicorn[standard]>=0.24.0

# HTTP client for API calls
aiohttp>=3.9.0

# Environment and configuration
python-dotenv>=1.0.0
pydantic>=2.5.0
pyyaml>=6.0

# AI and agents
openai>=1.3.0
openai-agents

# Stock data
yfinance>=0.2.20

# Background job scheduling
apscheduler>=3.10.4

# Development and testing
pytest>=7.4.0
pytest-asyncio>=0.21.0
pytest-cov>=4.1.0
pytest-mock>=3.11.1
httpx>=0.25.0
